Make code refactor by adding app.rb file
