+++ main.rb
@@ -1,18 +1,11 @@
-require_relative 'person'
-require_relative 'student'
-require_relative 'book'
-require_relative 'classroom'
-require_relative 'teacher'
-require_relative 'rental'
+require_relative 'app'
 
-class App
+class Main
   def initialize
-    @people = []
-    @books = []
-    @rentals = []
+    @app = App.new
   end
 
-  def start
+  def run
     puts '*************AUCA School Library Management************'
     puts '             ==============================          '
     puts 'Welcome to our School Library!'
@@ -21,156 +14,23 @@ class App
       option = gets.chomp
       break if option == '7'
 
-      option_output option
+      @app.option_output option
     end
-    puts 'Thanks for using our system!'
+    puts 'Thank you for using our Library!'
   end
 
   def menu
     puts
-    puts 'Please choose an option by entering a number: '
-    puts '    1 - List all books'
-    puts '    2 - List all people'
-    puts '    3 - Create a person'
-    puts '    4 - Create a book'
-    puts '    5 - Create a rental'
-    puts '    6 - List all rentals for a given person id'
-    puts '    7 - Exit'
-    puts '**********************************************'
-  end
-
-  def option_output(option)
-    case option
-    when '1'
-      list_all_books
-    when '2'
-      list_all_people
-    when '3'
-      create_person
-    when '4'
-      create_book
-    when '5'
-      create_rental
-    when '6'
-      list_rentals_by_person_id
-    else
-      puts 'Please enter a number between 1 and 7.'
-    end
-  end
-
-  def list_all_books
-    puts 'No books in the database! Please add a book.' if @books.empty?
-
-    @books.each { |book| puts "Title: #{book.title}, Author: #{book.author}" }
-    sleep 1
-  end
-
-  def list_all_people
-    puts 'No people in the database! Please add a person.' if @people.empty?
-    @people.map { |person| puts "[#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}" }
-    sleep 1
-  end
-
-  def create_person
-    print 'Would you like to create a student(1) or a teacher(2) [Please input a number]: '
-    option = gets.chomp
-
-    case option
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    else
-      puts 'Invalid input. Please input 1 or 2'
-    end
-  end
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp.to_i
-
-    print 'Name: '
-    name = gets.chomp
-
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = gets.chomp.downcase
-
-    student = Student.new(@class, age, name, parent_permission)
-    @people << student
-
-    puts 'Student created successfully'
-    sleep 1
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-
-    print 'Name: '
-    name = gets.chomp
-
-    print 'Specialization: '
-    specialization = gets.chomp
-
-    teacher = Teacher.new(age, name, specialization)
-    @people << teacher
-
-    puts 'Teacher created successfully'
-    sleep 1
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-
-    print 'Author: '
-    author = gets.chomp
-
-    book = Book.new(title, author)
-    @books << book
-
-    puts 'Book added successfully'
-    sleep 1
+    puts 'Please choose an option by entering a number'
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all rentals for a given person id'
+    puts '7 - Exit'
   end
 end
 
-def create_rental
-  puts 'Select a book from the following list by number'
-  @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title}, Author: #{book.author}" }
-
-  book_id = gets.chomp.to_i
-
-  puts 'Select a person from the following list by number (not id)'
-  @people.each_with_index do |person, index|
-    puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-  end
-
-  person_id = gets.chomp.to_i
-
-  print 'Date: '
-  date = gets.chomp.to_s
-
-  rental = Rental.new(date, @books[book_id], @people[person_id])
-  @rentals << rental
-
-  puts 'Rental created successfully'
-  sleep 1
-end
-
-def list_rentals_by_person_id
-  print 'ID of person: '
-  id = gets.chomp.to_i
-
-  puts 'Rentals:'
-  @rentals.each do |rental|
-    puts "Date: #{rental.date}, Book '#{rental.book.title}' by #{rental.book.author}" if rental.person.id == id
-  end
-  sleep 1
-end
-
-def main
-  app = App.new
-  app.start
-end
-
-main
+execute_app = Main.new
+execute_app.run
